Add setup teardown to integration tests
This adds an init and a teardown to the database for every test.
It fixes this error I encountered running the integration tests.

=== RUN   TestTimetableModelFetchAll
--- FAIL: TestTimetableModelFetchAll (0.00s)
    database_test.go:101: expected timetable key to be 'key1' was  this

This is because the test assumed the list of timetables would have
one entry(clean test environment), but the first element in the
array was acutally the timetable from the previous test.